Factor out codegen backend
